Widen predicate interface to VerificationContext and add Predicate protocol
Predicates now receive the full VerificationContext instead of three
separate args (agent_output, pre_env, post_env). This lets combinators
(all_of, any_of, not) forward observations and function_calls to nested
predicates unchanged — unblocking OCSF predicates that need to read
ctx.observations["openshell"] instead of smuggling data through env fields.

Predicate is now a Protocol in midojo.verifiers — any object with
evaluate(ctx: VerificationContext) -> bool satisfies it, replacing the
union type that required updating for every new predicate class.

The verifier framework (protocols, combinators, registry) moves from
midojo.verifier to midojo.verifiers.__init__; midojo.verifier becomes a
re-export shim for backwards compatibility.
